refactor: centralize component constants and types
