+++ src/@types/components/appToaster.ts
@@ -1,0 +1,11 @@
+export type AppToasterPosition =
+  | 'top-left'
+  | 'top-center'
+  | 'top-right'
+  | 'bottom-left'
+  | 'bottom-center'
+  | 'bottom-right';
+
+export interface AppToasterProps {
+  position?: AppToasterPosition;
+}
