+++ src/@types/components/cpu.ts
@@ -1,0 +1,5 @@
+export type RgbColor = {
+  r: number;
+  g: number;
+  b: number;
+};
