+++ src/@types/components/navigationDrawer.ts
@@ -1,0 +1,12 @@
+import type { ReactNode } from 'react';
+
+export interface NavigationDrawerProps {
+  open: boolean;
+  onClose: () => void;
+}
+
+export interface NavigationItem {
+  text: string;
+  icon: ReactNode;
+  path: string;
+}
