+++ src/@types/components/themeToggle.ts
@@ -1,0 +1,6 @@
+import type { SxProps, Theme } from '@mui/material';
+
+export interface ThemeToggleProps {
+  fixed?: boolean;
+  sx?: SxProps<Theme>;
+}
