+++ src/components/NavigationDrawer.tsx
@@ -9,27 +9,10 @@ import {
   Toolbar,
 } from '@mui/material';
 import React from 'react';
-import { BiHistory } from 'react-icons/bi';
-import { FaShare } from 'react-icons/fa';
-import { FiUsers } from 'react-icons/fi';
-import { MdClose, MdSpaceDashboard } from 'react-icons/md';
-import { RiSettings3Fill } from 'react-icons/ri';
+import { MdClose } from 'react-icons/md';
 import { Link } from 'react-router-dom';
-
-interface NavigationDrawerProps {
-  open: boolean;
-  onClose: () => void;
-}
-
-const drawerWidth = 200;
-
-const navItems = [
-  { text: 'داشبورد', icon: <MdSpaceDashboard />, path: '/dashboard' },
-  { text: 'کاربران', icon: <FiUsers />, path: '/users' },
-  { text: 'تاریخچه', icon: <BiHistory />, path: '/history' },
-  { text: 'اشتراک گذاری', icon: <FaShare />, path: '/share' },
-  { text: 'تنظیمات', icon: <RiSettings3Fill />, path: '/settings' },
-];
+import type { NavigationDrawerProps } from '../@types/components/navigationDrawer';
+import { drawerWidth, navItems } from '../constants/components/navigationDrawer';
 
 const NavigationDrawer: React.FC<NavigationDrawerProps> = ({
   open,
