+++ src/constants/components/network.ts
@@ -1,0 +1,1 @@
+export const MAX_HISTORY_MS = 90 * 1000; // 1 minute 30 seconds
